RIP roscoring, it was fun while it lasted.
So I was compiling my notes on score calculation when I stumbled upon something interesting. As it turns out, there already IS an optimal gear calculation tool - the Gear Calculator.
https://docs.google.com/spreadsheets/d/1Ch_RsrgIOjat7WKMeSxKwTPGq0yA8YUaKfV17u0sdic
While it doesn't automatically optimize gear like my program, that task isn't very tough for a human. There's even a complete dataset of gears, minis, song info, hell, it's got per-second NPS data for every song in the game. I can't compete with that!
So it seems my days as an archivist of sorts is over. It's been quite a journey, this repository is the only reason I didn't quit Robeats entirely after 100%ing the song library. Oh well! I had some fun making and maintaining this thing. Now that there's no reason for it to exist, it's time to move on. Like half of the Robeats community has already done that anyway.